Updated HTTP headers model to be a struct
Necessary to provide safe defaults and relax having all attributes filled so you can make non-device API calls. This also allows the device to not always have to provide all information too (which is sometimes the case).

Reek was updated to ignore the multiple initialize statements because, in this case, there's no quick way to break this down further.

Milestone: patch