Fixed API setup endpoint
Necessary to handle the registration of new devices by doing the following:

* Setting a default device label.
* Generating a Friendly ID for the device.
* Saving the provided MAC address.
* Generating an API key.
* Saving the current time for setup date/time.

Otherwise, an existing device will be answered.

This also makes use of the new setup BMP image which is now part of our assets.

In summary, all of this means you can automatically register a new device once connected to your network.

Milestone: patch