Refactored health slice app assets as main assets
Necessary to add clarity to the _kind_ of assets being imported into the Health slice. The concept, moving forward, is that the main/primary app (non-slice) will be known as `main` and then if any components need to be shared between the slices, those will use each slice's name as the prefix.

Milestone: patch