Updated device actions to detect htmx requests
Necessary to render the views with or without htmx support for graceful degradation when JavaScript isn't enabled. This is also being done to support viewing device logs since navigating between device information and log informnation would be too cumbersome without these changes.

The logic for detecting htmx requests is duplicated throughout the controllers, for now, with the intent to remove this duplication in the future.

Milestone: minor