Added API client
Necessary to provide a low-level, configurable, monadic client for making API requests. This builds upon the configuration (added earlier). At the moment, this only knows how to make HTTP GET requests but the idea is you could expand this to support other HTTP verbs.

This will soon be used to talk to the Core server.

Milestone: minor