Added screen poller
This object builds upon all of the work, thus far, by allowing a device to proxy all images to the Core server for local display. The associated script will pull Core's Display API every five minutes (this is the default but can be customized).

This is meant to be added to your `Procfile` so the polling can run as a background job. The project documentation for this will be updated soon.

While simple and effective, this isn't meant to be the final solution. Only a stop gap to tie us by while I work out the a better implementation because all images are stored on the file system but there is no distinction between what device owns which image. Plus, we also load images by oldest modified time and are entirely dependenant on the file system (which may or may not be a good idea).

Code coverage has been disabled, temporarily, for the injected dependencies since testing the full stack requires the remote server which isn't ready yet.

Milestone: minor