Refactored use of access token as API key
Necessary to use consistent terminology throughout the objects.

From an HTTP API standpoint, we use _access token_ but internally, we use _API key_. Both are the same for different names used in different contexts. This helps make that distinction more clear and improve consistency. This also matches up with how this information is stored in the database since we use _API key_ for all device records.

Milestone: patch