Added lib container spec
Normally, you wouldn't test this container since this container is indirectly tested in all objects that inject this container's dependencies. That said, and with the introduction of the RSpec shared context for this container, SimpleCov doesn't see our `HTTP` client being used so thinks there isn't code coverage around `HTTP`. This addresses that concern.

Milestone: patch