Added editor page
Necessary to modify source code in real time and render the generated image as a preview using our middleware which produces event streams for Server Side Rendering (SSE). This appears as a preview image. The editor is where the HTML/CSS code goes for generating the image. The editor triggers new image generation upon page load _and_ when changes are maded to the text area (via every keyup event with a slight delay in order to not overload the server).

Each preview image is uniquely identified by timestamp.

Issue: 5
Milestone: minor