Added editor create action
Necessary to handle the creation of a new screen based on incoming template information. For the moment, all processing is handled by htmx POST requests. There is potential for updating the page when using normal form submission which is why the show view is used in that case. This will need to be expanded upon further soon.

Issue: 5
Milestone: minor