Removed the development container
This is being removed in favor of using the new Docker Compose configuration since it's more universal than using the `.devcontainer` which is more specific to VSCode (although there is generic support for other editors but not as widely used).

This also reduces the maintenance burden of not having to maintain multiple Docker files.

Issue: 57
Milestone: minor