Fixed site footer assets and links
Necessary to extract inline SVG images in favor of using SVG images as now stored in the assets `icons` folder. The footer links were cleaned up to only link to the GitHub source once (with a preference for the social icon). The Ruby and Hanami version information has been removed and all links aligned at the end of the footer (i.e. `flex-end`) since this information isn't as relevant now that we have a better looking footer.

Milestone: patch