Fixed screen downloader to apply proper file extension
Image's provided by the core server are unreliable in that they sometimes have file extensions and sometimes don't. In order to ensure all downloaded images have the correct file extension, we use MiniMagick to obtain the correct MIME type and then rename the downloaded file as such. This has a few benefits:

* We visually see the type of image when looking in the assets directory.
* We ensure, when serving images to the device, that the device can properly render the image because the device also can't figure out the type of image without the file extension.

As an added safety check, we'll answer a failure if there is potential file corruption when MiniMagick can't read the file.

Milestone: patch