Fixed setup script to build common environment and not compile assets
Necessary to ensure the _common_ environment (i.e. `.env`) is setup properly for all environments because we need the `PG_*` variables for both development and production purposes (especially when using Docker).

Asset compilation has been removed in favor of letting Overmind handle this because `hanami assets compile` only compiles for production, not development. There is a `--env development` option you can pass to the CLI but doesn't work.

Milestone: patch