Fixed information icon dimensions and rendering
Necessary to help with page rendering and load performance. For the devices page, we now leverage our information SVG icon (instead of keyboard emoji) which is consistent with how we are using information sources elsewhere in the UI.

The CSS was updated to a mouse pointer and flex display which feels more natural and intuitive.

Milestone: patch