Added device and device log pages search query retainment
Further enhances searches by ensuring the search query is populated in the search form (this is passed down via the view). This is especially handy when you've bookmarked the page.

The views and templates have been updated to show empty results (by using the shared empty partial) when there are no records. This works when dynamically searching or loading from a bookmarked link.

Milestone: minor