Updated Reek configuration with excluded actions
These issues are of the same ilk related to feature envy with `request` and `parameters`. This issue would go away if we had direct access to these objects within each action and is the root cause of having to constantly pluck out these objects. In the meantime, until I can fix Hanami further, I've added these actions to the configuration.

Milestone: minor