Updated screens creator to use keyword arguments
Necessary to simplify the method parameters so you don't have to construct a hash as the first positional argument. Instead, you pass in the output path as the first positional argument and then double splat the keyword arguments, as desired, to configure behavior.

This makes the Object API less awkward and improves usage from within the console too.

Milestone: minor