Updated API device contract as schema
Necessary to make room for schemas (Dry Schema) and contracts (Dry Validation). This will allow both types to coexist and build upon one another instead of having a single contracts namespace for both. This also includes the new attributes for dealing with device sleep range:

- `sleep_start_at`
- `sleep_end_at`

The RSpec helper was updated to account for both schemas and contracts as well. This includes having all namespaces alphabetized for quicker lookup. The corresponding spec was updated to improve readabilty by using the subject since it was defined but never used but now is.

Issue: 109
Milestone: minor