Fixed Style/ItBlockParameter issue
Necessary to avoid using `it` in a multi-line block. This improves readability but Reek doesn't know how to check `it` so now detects feature envy which was there to being with. Marking this as ignored because Reek is aggressive in this regard plus this is a flaw of the `Socket` Object API for not making this easier to work with.

Milestone: patch