Added Docker Compose application and PostrgreSQL ports
Necessary to define these ports via the global `.env` file for use within the production and development compose configuration. This includes refactoring the compose files so the keys have the same structure and order for improved readability and comparison.

This also removes the `env_file` keys from the template since we don't need to be explicit since this'll be implicitly picked up.

Milestone: minor