Added device repository mirror playlist
Necessary to update all devices to the same playlist ID (or unset) if desired. In other words, being able to _mirror_ the same playlist across multiple devices. This uses a SQL CASE statement which applies the `playlist_id` for all device IDs listed and nullifies the devices not identified.

Milestone: minor