Fixed Dockerfile creation of assets folder
This cropped up do to cd5188bf4855 (Removed Docker screen assets) because I forgot that we still must create the assets folder. We got this for free originally, because we were always creating sub-folders but now that the subfolders are gone, we still must create the root folder.

Issue: 186
Milestone: patch