Added Display API error screens
Necessary to improve the user experience by providing more feedback for non-fatal errors that we can catch and render an error image for. This uses the newly added screen gaffer to render a specific error screen for the device with feedback on how to take corrective action. This allows folks to quickly resolve the issue without having to sift through server logs or hitting the Display API endpoint for further information.

Issue: 191
Milestone: minor