Added custom view context
Necessary to determine if the HTTP request is an htmx request. The `#htmx?` method is a convenience method around the HTTP request object and, for testing purposes, must be set via the instance variable since Hanami makes it very difficult to set the request object once Hanami is loaded.

This will soon be used within the view layer to determine if the htmx is being used or not.

Issue: 209
Milestone: minor