Updated shared title partial to include title tag when htmx is used
Necessary to ensure the HTML `title` tag is rendered when the HTTP request is a htmx request so we can let htmx update the page title when rendering HTML fragments. Otherwise, the previous page's title is used which is not desired. Here's how to think about this:

* *Page*: Uses `content_for` to set the page title for a full page render (i.e. `layout: true`).
* *Fragment*: Uses the tag helper to render an HTML `title` element (i.e. `layout: false`).

Issue: 209
Milestone: minor