Added screen struct image replace behavior
Necessary to provide a way to replace an existing image attachment. This needs more work because Shrine has a similar capability but requires clearing of the cache from time to time so using this solution, for now, as a temporary solution.

This also teaches the `#attach` and `#upload` methods to answer themselves to match parity with the new `#replace` method. The specs were updated accordingly which improves consistency.

Milestone: minor