Updated RSpec helper to ignore Shrine warnings
Necessary to clean up the test suite output by ignoring the Shrine warnings since the maintainer won't fix the issue. This is definitely a concern since use of `URI::RFC3986_PARSER.make_regexp` will not be supported in the future but, for now, we can ignore this noise in test suite so we can detect any issues with our own code sooner rather than later (which is more important).

Milestone: patch