Added screens compressor
Necessary to compress BMPs as PNGs. Otherwise, when already a PNG, this works as a passthrough. No additional image quality settings are needed as ImageMagick's default settings will handle BMP to PNG conversion with the correct amount of compression.

This was discovered, when synchronizing images from Core, where the BMP image downloads would be 1 MB or more. Our devices will not display images that large. Ultimately, this is a bug with the Core server so, instead of trusting Core, we'll compress the image ourselves when a BMP image is detected.

This will soon be folded in the screens synchronizer.

Reek has been updated to ignore feature envy and too many statements since this is an aspects of the MiniMagick Object API and breaking this down further is a limitation of MiniMagick in this case.

Milestone: minor