Added site navigation and menu styles for small screens
Necessary to provide menu styling for smaller screens so that the menu collapses down into a hamburger icon which, upon clicking, will expand to a list of menu items. This is done by styling a checkbox to show and hide this list based on whether the checkbox is clicked or not. The checkbox is always hidden but allows for the dynamic effect of showing/hidding the menu. CSS transitions are used to turn the hambuger into an "X" so you have a visual clue on how to close the menu.

Milestone: minor