Fixed screen designer middleware to use Nginx compatible headers
When using an Nginx reverse proxy in front of this application, use of the `Connection` header with keep alive doesn't play nice with Nginx. Instead, that header needs to be dropped and the custom `X-Accel-Buffering` header used instead. See this link:https://serverfault.com/questions/801628/for-server-sent-events-sse-what-nginx-proxy-configuration-is-appropriate[discussion] for details.

Milestone: patch