Added playlist part
Necessary to render current screen URI (if associated). Unfortunately, ROM Factory can't build nested associations. Worse, you can't stub the associations either. So the spec setup is more complicated than desired but ensures a part is iniitialized with a playlist that has a current item.

All functionality is focused the current item of the playlist which can be a pill that displays current status along with the current screen URI (if present).

Milestone: minor