Fixed search styles for smaller resolutions
Necessary to allow the search to wrap and remain centered (both horizontally and veritically) on smaller displays. The search and clear actions are wrapped now. This includes removing top and bottom margins from the buttons which will be cleaned up further after this.

The CSS class was renamed to `bit-search-form` in order to clarify what this style is meant for because the form is what is being styled even though it is wrapped in a `search` element. This allows room for styling the `search` element in the future if desired.

Milestone: patch