Added screen struct dated image name
Discovered our Firmware only renders an image if the image name has changed. Doesn't matter if the content is different or not. This is a problem when making updates to the same image where the file name never changes especially when using the Screens API and patching/updating an existing image.

To counteract this situation (and provide a form of Firmware specific cache busting) a date/time stamp (example: `Time.now.utc.to_i`) is added as a suffix. This will ensure -- and only when the image is updated -- to force the Firmware to replace the image. This still respects the original intent of the Firmware code while allowing updated images to be properly refreshed.

Issue: 213
Milestone: patch