Added screen download link
Provides a convenient and quick way to download a screen.

Unfortunately, the `link_to` helper can't be used because Hanami/Tilt causes spec failures due to not understanding the `download` attribute. This only fails in the test environment which is why a standard HTML `a` element is used instead.

Milestone: minor