Updated screens color mapper to handle bit depths higher than four
Necessary to allow higher bit depths by logging a debug message when a bit depth highter than four is given. This'll allow for the various converts to handle accordingly by checking if the path exists or not.

RuboCop and Reek have been updated since there breaking down the methods won't help in this case.

Milestone: minor