Updated screen creators and converters to use mold
Necessary to leverage the new mold struct which houses all attributes for creating and converting screens into the appropriate format necessary for rendering on a device. This has the added benefit of abstracting away the model so we can use attributes that are not model specific.

RuboCop was updated to ignore the performance concern since we are updating a `Struct` and not a `Hash` and RuboCop can't differentiate when using a refined mold.

Issue: 219
Milestone: patch