Updated screens create and patch actions to accept screen parameter
Necessary to ensure consistency in API design so the endpoint you are making requests to defaults to the same top level key required in the request payload. This reduces confusion and avoids the awkward use of the original `image` top-level key.

This is a _breaking change_ but posted about this change in `#byos` Discord last week.

Milestone: minor