Added image uploader checksum calculation
Necessary to start calculating checksums for all image attachments so we can determine if they've changed or not. This _is a breaking change_ for existing specs that will be fixed in the next commit.

RuboCop has been updated to allow `md5` since this problem stems from Shrine and not us.

Issue: 213
Milestone: minor