Updated screen struct to calcumate image name with checksum
Necessary to ensure the screen file name includes the new checksum which only changes when the attached image is different. This ensures the device will ignore an image that hasn't changed since the `updated_at` timestamp wasn't accurate enough.

The method name has been changed to reflect this new behavior. This is also a _breaking change_ that will be addressed in the next few commits.

Issue: 213
Milestone: minor