Updated screen encoder to use image file name with checksum
Necessary to provide more integrity for device caching purposes. The checksum will be used instead of the updated at timestamp to ensure a new image is only displayed when the actual image is different.

The associated screen synchronizer and display specs have been updated to account for this change as well.

Issue: 213
Milestone: minor