Updated screen shoter to merge browser settings and options
Necessary to incorporate the newly added application browser settings by injecting them and merging them with the browser options. This also ensures that when a browser instance isn't created in time, a process timeout error will be thrown while avoiding the quitting of the instance since it doesn't exist yet.

Issue: 222
Milestone: minor