Fixed browser setting defaults
Necesssary to provide safe defaults and ensure they show up in the error responses. The defaults are done at the type level rather than the settings level which is why the setting only uses an empty hash as the default.

Support for integer has been added so you can use an integer or float for the `process_timeout` and `timeout` values. This also ensures the defaults are supplied when only partial settings are supplied.

Issue: 225
Milestone: patch