Added delete shared partial
Necessary to extract common functionality for deleting records. By default, this is focused on single record deletion which is why fetching of optional local keys (with safe fallbacks) is used. Otherwise, for bulk deletion, you can supply a specific label and class. This'll soon be used to refactor existing templates and partials to reduce duplication.

Milestone: minor