Added HTTP model header processing for Display API requests
Necessary to parse and validate the new `HTTP_MODEL` header supplied by the Firmware when processing Display API requests. We can't process this information, as of yet, because the provided value is `og` instead of `og_png` or `og_plus`. This means we don't have a valid way to determine if we are communicating with a 1-bit or 2-bit device. I've logged a bug with the Firmware team for further action: https://github.com/usetrmnl/trmnl-firmware/issues/234.

Milestone: minor