Added authentication slice feature
Necessary to tap into Rodauth's plugin feature by creating a Hanami specific feature which works as the translation layer between Rodauth and Hanami. In this case, we are swapping out Rodauth's view rendering with Hanami's view rendering but only if one of our Hanami templates exist. Otherwise, we fall back to Rodauth.

RuboCop was updated to ignore the block length because, for now, it's best to keep all of this logic in one place without splitting this up further.

Milestone: minor