Added application action authorize callback
Necessary to ensure Rodauth is used to authenticate all requests.

Only when Rodauth isn't provided do we return early because we are unit testing actions which don't use routes which is also when Rodauth will be `nil`.

Reek has been updated to ignore the mulitple statement since there isn't a great way to reduce this further. SimpleCov is being ignored because the code is tested but I think SimpleCov is confused by the throw and catch logic.

Milestone: minor