Updated to Caliber 0.84.0
link:https://alchemists.io/projects/caliber/versions/0.84.0/[Details]. This also picks up a Rouge patch.

One of the interesting quirks of this change is that the RuboCop team changed how `it` works for functional pipelines. I've refactored the code to account for this change but also raised concern with the RuboCop team since this was surprising to introducing in a minor version bump.

Milestone: minor