Refactored RSpec specs to use user statuses shared context
Necessary to leverage the newly added shared context, remove the need for the database seeds CLI command, and reduce the verbosity of the test suite output due to using the original CLI command.

The Hanami helper was updated to automatically include the user statuses since they are needed for feature and request specs.

Milestone: patch