Added user part
Necessary to capture standard user attributes plus the `password` attribute which behaves like a virtual attribute for form input (Reek was updated to ignore this since we can't avoid it). The pill method is also introduced to provide CSS pill class for denoting different users statuses within the UI (this'll soon be used for the user page).

Milestone: minor