Fixed Puma configuration to use less aggressive concurrency
The original thought was to let link:https://github.com/ruby-concurrency/concurrent-ruby[Concurrent Ruby] calculate how wide to go with concurrency based on processor count but this can mean that more resources are allocated that might not be desired for simple single server situations. This simplifies the implementation further by removing the Concurrent Ruby dependency and relying on manually adjusting `HANAMI_WEB_CONCURRENCY` for more concurrency when desired.

Milestone: patch