Added Alpine.js
The overarching goal of this project is to keep JavaScript to a minimum and use modern HTML and CSS as much as possible but there are times where handling small UI tweaks is best handled with libraries like link:https://alpinejs.dev[Aline.js]. This will soon be used to enhance the UI further with small DOM manipulations.

The application layout has been updated to defer loading of JavaScript since Alpine.js requires this or you'll end up where in the web console where the JavaScript was loaded before the `body` tag is ready. This has no effect the htmx library.

Milestone: minor