Added visibilty button shared partial
Necessary to dynamically render a button with either an eye icon to show information or a disabled eye icon to hide information using Alpine.js to toggle between the two states. This is meant to be used within an input group where `show` value is initialized as `false`. Example:

[source,erb]
----
<div class="input-group" x-data="{ show: false }">
  <%= render "shared/visibility_button" %>
</div>
----

Milestone: minor