Fixed Rodoauth JWT access token period setting and configuration
A couple things went wrong here for which I'm not sure why I missed this originally. The first is that the setting was not properly converting the environment value's string into an integer. The second was that I wasn't using the right Rodauth configuration to ensure the `api_access_token_period` setting properly mapped to the to Rodauth `jwt_access_token_period` configuration.

With this change, you can properly adjust the amount of time a JWT access token will live.

Milestone: patch