Added version information to footer
Necessary for debugging purposes so people can clearly see (and link to) the latest tag and/or SHA. When using a specific version, the tag and SHA will be associated but if running off of the `main` branch, the commit SHA will, usually, be farther ahead than the tag. In either case, this information will help determine what version of the code base the application is running on.

Milestone: minor