Removed firmware version from welcome screens
Necessary to reduce confusion when creating a virtual device and seeing a question mark for firmware version (since it's unknown) and then later -- after getting a firmware update -- you don't see the firmware version reflected in the welcome screen. Opted to remove this information since this is provided by the firmware now and you can also find this information when viewing a device (which is also more dynamic and up-to-date). This is better than trying to keep this screen up-to-date since it's meant to be temporary.

Milestone: minor