Added extension migration
Necessary to manage extension records which are an evolution of what is known as _plugins_ in Core. The design of this table combines Core plugin types such as native, private, public, third party, and recipes. The column names and terminology is simplified and adapted for this architecture in order to avoid the mistakes of the Core system while also being able to scale and support different extensions in the future.

Milestone: minor