Added boolean schema coercer
Necessary to provide a reusable boolean coercer that can be used for multiple schemas (not only devices). This is an extraction and simplification of the existing `DeviceValue` coercer that will soon be replaced with this function instead. This is also more customizable since you provide the key, who's value needs to be coerced, as the first argument. This also allows this function to be curried which improves reusablity across schemas.

Milestone: minor