Added logger provider
Necessary to avoid the default logger and replace with an advanced logger. Hanami will automatically check if a custom logger provider exists and, when it does, use the custom provider (this provider) instead. Custom filters, specific to this application, are applied to ensure sensitive information isn't leaked. Otherwise, only a JSON logger is used for production purposes while a color logger with JSON logging is used for all other environments. link:https://alchemists.io/projects/cogger[Documentation].

SimpleCov is disabled for the environment logic because testing a provider is not easily supported in Hanami. This will fall out in normal usage anyway.

Milestone: minor