Updated RSpec helper to ignore Hanami Rack logger warnings
Necessary to prevent the overwhelming spamming of warnings to the console when running the app and test suite due to Hanami failing to be diligent in resolving dependency deprecation warnings. I've logged an issue with the Hanami team that has been fixed but unsure when, or how soon, a new version release will be published.

Milestone: patch