Added htmx configuration to context
Necessary to provide a default configuration while allowing for additional customization if desired. Customization is made possible by using the `content_for` method with the `htmx_merge` key. The value can be any hash that adds or overrides attributes.

This cropped up because you can't use "textContent" as a swap style when using the Server Sent Events (SSE) extension when the global default swap configuration is "outerHTML". I've logged a bug with the htmx team but doubt this'll be resolved anytime soon given how slow existing issues are dealt with.

Milestone: minor