Fixed designer event stream spec to remove unused payload variable
Necessary to resolve the following warnings:

----
warning: assigned but unused variable - payload
----

The reason this popped up was because we need to message `#each` in order to verify log output but we don't want to do anything with the block. We also can't have a block that does nothing, otherwise we'd get a RuboCop Lint/Void error so we do a boolean check instead (which also does nothing but satisfies all checks).

Milestone: patch