Updated to Hanami 2.3.1
Necessary pick up the patch for the circular router Ruby warning. This was an issue I logged with the team and is now resolved.

This also picks up ERB 6.0.0 which doesn't appear to break anything and seems to be in preparation for Ruby 4.0.0 this Christmas. link:https://github.com/ruby/erb/releases/tag/v6.0.0[Details].

Milestone: patch