Added Valkey keyvalue service
Necessary to support an open source solution with better licensing. link:https://valkey.io[Valkey] is now the default keyvalue data store.

Use of `keyvalue` instead of `redis` is done to be more generic. This also allows folks to switch to Dragonfly and/or Redis if desired.

Docker was updated to refer to this as the _keyvalue_ service instead of the _cache_ service.

Issue: 243
Milestone: minor