Added liquid json filter
Without an explicit json filter, it was not possible to directly pass an entire JSON object from a fetched URL to the template in an extension. The worker would die with: `SyntaxError: Unexpected token '=>'`.

Use of `JSON.generate` instead of using the `JSON` function is important because it'll cause a `Liquid error: internal` error when casting a string.

Reek was updated to ignore the utility functionality of this implementation because this stems from Liquid's design and bleeds into our architecture.

Co-authored-by: Brooke Kuhlmann <brooke@alchemists.io>
Milestone: minor