Fixed extension fetcher MIME type processing
Ensures resources -- with an incorrect MIME Type -- are parsed as desired when the extension's content type header is supplied. When no header is specified for the extension, the MIME Type will fall back to the response's MIME Type.

This is useful for working around servers, like GitHub, that notoriously use plain text for all raw content (including gists) despite the file extension providing the correct format.

Issue: 253
Milestone: patch