Fixed extension HTTP headers to use accept instead of content type
This was a mistake of work applied in cc327ba5ac0f (Fixed extension fetcher MIME type processing) where the `Content-Type` header was incorrectly used. Instead, the `Accept` header should be used because that's what this link:https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Accept[header] was meant for. With extensions, we take this a step further by, first, allowing the server to honor our request and, if not, we'll format the response ourselves as a fallback.

Issue: 253
Milestone: minor