Updated playlist screen show action to advance screen
Necessary to ensure the current screen is updated when advancing the playlist but only if the playlist is in _manual_ mode.

This is a slight cheat since we're not using an HTTP PATCH request to update the playlist record but this keeps us from having to use a form. Additionally, even if we had a form, there's nothing to provide as we have all the information we need to update the playlist accordingly.

Reek has been updated to ignore the feature envy issue with referencing the `request` object too many times in the `view_settings` method. Not much we can do about this since Hanami doesn't provide much of an Object API for this.

Milestone: minor