Updated playlist screens index action to load first or current screen
Necessary to ensure the current screen (if set) is honored when loading screens. This means if the current screen is present (via the current time), the first slide loaded will be the current screen. Otherwise, the first screeen will be used. If all else fails, no screens will be displayed as a default message.

At the moment, only this action needs this functionality but `.load_screens` should probably be extracted which is why the Reek comment helps call attention to this.

Milestone: minor