Fixed popover content partial headings
Dialogs can have H1 elements which don't break the semantic outline of the DOM so these are clearly marked as such. The style was updated to ensure the font size is explicitly set.

Sub-headings were updated to be H2 instead of H3 so the hierarchy within the dialog makes logical sense.

Milestone: patch